added Scheduler which deletes expired tokens from DB

+++ src/main/java/com/nikki/jwt/security/repository/SecurityUserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface SecurityUserRepository extends JpaRepository<SecurityUser, Integer> {
+public interface SecurityUserRepository extends JpaRepository<SecurityUser, Long> {
 
     Optional<SecurityUser> findByEmail(String email);
 }
